+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Date Form element (#520)
 - Update refresh for dashboard scene (#522)
 - Updated Text area, Code editor elements to escape multi-lines (#520)
+- Update refresh for dashboard scene using useDashboardRefresh hook (#528)
 
 ## 4.7.0 (2024-10-08)